Fixed all the routing problems with the pages.
