+++ cracking_the_resume/seeds.js
@@ -1,6 +1,6 @@
 var mongoose = require("mongoose");
 var User = require("./Models/user");
-var Resume = require("./Models/resume"); 
+var Resume = require("./Models/resume");
 
 //sample user data 
 var userData = [
@@ -37,18 +37,18 @@ var userData = [
 ]
 
 
-//sample resume data 
-var resumeData = [
-	{
-		resumeName: "r1.pdf"
-	}, 
-	{
-		resumeName: "r2.pdf"
-	},
-	{
-		resumeName: "r3.pdf"
-	},
-]
+// //sample resume data
+// var resumeData = [
+// 	{
+// 		resumeName: "r1.pdf"
+// 	},
+// 	{
+// 		resumeName: "r2.pdf"
+// 	},
+// 	{
+// 		resumeName: "r3.pdf"
+// 	},
+// ]
 
 
 function seedDB(){
@@ -71,44 +71,36 @@ function seedDB(){
 	});
 
 	 
-		//creating resume records 
-	resumeData.forEach(function(seed){
-		//creating resume record 
-		var resumeRecord = new Resume(seed)
-
-		resumeRecord.save(function(err, resume){
-			if(err){
-				console.log("There was an error in saving the resume"); 
-			}
-			else{
-				console.log("Sucessfully created resume record"); 
-				console.log(resume);  
-			}
-		});
-	}); 
-
-	// userData.forEach(function(seed){
-	// 	User.create(seed, function(err, data){
+	// 	//creating resume records
+	// resumeData.forEach(function(seed){
+	// 	//creating resume record
+	// 	var resumeRecord = new Resume(seed)
+    //
+	// 	resumeRecord.save(function(err, resume){
 	// 		if(err){
-	// 			console.log("There was an error");
-	// 			console.log(err); 
+	// 			console.log("There was an error in saving the resume");
 	// 		}
 	// 		else{
-	// 			console.log("Added a new user!");
+	// 			console.log("Sucessfully created resume record");
+	// 			console.log(resume);
 	// 		}
-
 	// 	});
 	// });
 
-}
-Resume.find({}, function(err, resume){
-		if(err){
-			console.log("no resumes in the database");
-		}
-		console.log("this is the resume ids"); 
-		console.log(resume._id); 
-	})
+	userData.forEach(function(seed){
+		User.create(seed, function(err, data){
+			if(err){
+				console.log("There was an error");
+				console.log(err);
+			}
+			else{
+				console.log("Added a new user!");
+			}
 
+		});
+	});
+
+}
