+++ cracking_the_resume/views/signUp.ejs
@@ -14,7 +14,7 @@
     <div class="col-md-2">
     </div>
     <div class="col-md-8 conForm">
-            <form name="userData" action = "/split" class= "form-horizontal" role="form" method="post" enctype="multipart/form-data">
+            <form name="userData" action = "/redirect" class= "form-horizontal" role="form" method="post" enctype="multipart/form-data">
                 <h2>Let's get started</h2>
                 <div class="form-group">
                     <label for="firstName" class="col-sm-3 control-label">First Name</label>
